refactor test(challenge): challenge service 패키지 세분화
